requirement.txt
Pillow==4.2.1
pyperclip==1.5.27
qiniu==7.1.5